fof8-gen/pyproject.toml
[project]
name = "fof8-gen"
version = "0.1.0"
description = "A suite of tools for FOF8, starting with data automation."
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "opencv-python>=4.13.0.92",
    "pillow>=12.2.0",
    "pyautogui>=0.9.54",
    "pyyaml>=6.0.3",
]

[project.scripts]
gather-data = "fof8_gen.automation:main"

[build-system]
requires = ["hatchling>=1.27.0,<2"]
build-backend = "hatchling.build"

[tool.uv]
package = true
